Add generic arithmetic instruction emitter
This adds a new generic arithmetic instruction emitter, which can be
constructed using syntax like:
new InstructionEmitterArithmetic("SUB", InstructionSubtract::new,
InstructionSubtractImmediate::new).
This is to prepare for the new bitwise instructions.
