+++ src/main/java/li/cil/tis3d/system/module/execution/compiler/Compiler.java
@@ -2,11 +2,12 @@ package li.cil.tis3d.system.module.execution.compiler;
 
 import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableMap;
+
 import li.cil.tis3d.Constants;
 import li.cil.tis3d.Settings;
 import li.cil.tis3d.system.module.execution.MachineState;
 import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitter;
-import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitterAdd;
+import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitterArithmetic;
 import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitterJump;
 import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitterJumpEqualsZero;
 import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitterJumpGreaterThanZero;
@@ -18,9 +19,12 @@ import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmit
 import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitterNeg;
 import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitterNop;
 import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitterSave;
-import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitterSubtract;
 import li.cil.tis3d.system.module.execution.compiler.instruction.InstructionEmitterSwap;
 import li.cil.tis3d.system.module.execution.instruction.Instruction;
+import li.cil.tis3d.system.module.execution.instruction.InstructionAdd;
+import li.cil.tis3d.system.module.execution.instruction.InstructionAddImmediate;
+import li.cil.tis3d.system.module.execution.instruction.InstructionSubtract;
+import li.cil.tis3d.system.module.execution.instruction.InstructionSubtractImmediate;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -151,7 +155,7 @@ public final class Compiler {
     static {
         final ImmutableMap.Builder<String, InstructionEmitter> builder = ImmutableMap.<String, InstructionEmitter>builder();
 
-        addInstructionEmitter(builder, new InstructionEmitterAdd());
+        addInstructionEmitter(builder, new InstructionEmitterArithmetic("ADD", InstructionAdd::new, InstructionAddImmediate::new));
         addInstructionEmitter(builder, new InstructionEmitterJump());
         addInstructionEmitter(builder, new InstructionEmitterJumpEqualsZero());
         addInstructionEmitter(builder, new InstructionEmitterJumpGreaterThanZero());
@@ -162,7 +166,7 @@ public final class Compiler {
         addInstructionEmitter(builder, new InstructionEmitterNeg());
         addInstructionEmitter(builder, new InstructionEmitterNop());
         addInstructionEmitter(builder, new InstructionEmitterSave());
-        addInstructionEmitter(builder, new InstructionEmitterSubtract());
+        addInstructionEmitter(builder, new InstructionEmitterArithmetic("SUB", InstructionSubtract::new, InstructionSubtractImmediate::new));
         addInstructionEmitter(builder, new InstructionEmitterSwap());
 
         EMITTER_MAP = builder.build();
